extract media helper
